Added some help for users browsing for idea URLs
If a user tries to go to an idea that does not exist, it will tell them
that the idea doesn't exist instead of going to a blank idea page. If
the user is browsing an idea that gets deleted, the user will be
informed that the idea has been deleted.