HTTPS for Open Sans and Ordering Inject Index.html
Production was blocking the request for the Open Sans font because it
was over HTTP not HTTPS. Made that change.

Also the reordering of the damn index.html injections is pissing me off,
so here's hoping I fixed it.

One last thing, I specified commits for projects in the package.json
that didn't have Releases or Tags. To make sure they don't update the
package and break our stuff.