Revise event emissions
Event emissions are on a half second delay and fire the most recent
event in that half second. It seems that performance is hardly affected
with a mean of 199ms per request, median of 122ms, on 300 idea
submissions with up to 50 concurrent connections. Anything less than
500ms is probably pretty good considering that's the Google standard for
page rendering. Anything closer to 1 second will be noticeable by users.