Liking Ideas is an atomic operation
Removes concurrency issues around Liking ideas. However we need a valid
way of returning User data from the server back to the client. We cannot
query when we need it because ng-show, ng-hide, and ng-if query on the
digest cycle which is way too much for our servers to handle.