Adds new DB method
Needed to make an updateOnePullArray method for users.js and ideas.js to
use when liking/unliking an idea, since there was a slight difference
from how comments are stored in the DB. This should also help for
updates and backs in the future.

Also had to make the formats for ObjectIds match when pushing to the
array and pulling from it for ideas.